// Projects/Mobile/easyEnglishApp/models/types.ts
  export enum ExamTestType{
    IELTS,
    PTE,
    GE,
  }
  
  export enum ExamTestSectionType{
    Reading,
    Writing,
    Speaking ,
    Listening
  }